trie_rs: lift automatons into a top-level automaton module
The Automaton trait, StateClass, the byte-level wildcard NFA, and the
UTF-8-aware automatons (CaseFoldExact, the Levenshtein DP row and
bit-parallel NFA, CodepointWildcard) depend only on each other, not on
any trie type, yet lived under trie_map/iter and str_trie_map. Give
them a top-level trie_rs::automaton module.

The trie-coupled pieces stay with the iterators: the AutomatonIter
driver moves to trie_map/iter/driver.rs, and the WildcardIter
dispatcher and WildcardBackend join the filter iterator in
trie_map/iter/wildcard.rs.

Automaton docs now state the contract only; how the iterators drive an
automaton is documented on the iterators.
