+++ src/redisearch_rs/trie_rs/src/automaton/wildcard/atoms.rs
