+++ src/redisearch_rs/trie_rs/src/str_trie_map/mod.rs
@@ -9,7 +9,6 @@
 
 //! A UTF-8 keyed trie map. See [`StrTrieMap`].
 
-pub mod automaton;
 pub mod iter;
 
 use crate::TrieMap;
@@ -130,14 +129,14 @@ impl<Data> StrTrieMap<Data> {
 
     /// Yield every entry whose key equals `needle` after per-codepoint case
     /// folding, in lexicographical key order. See
-    /// [`CaseFoldExact`](automaton::CaseFoldExact) for the matching model.
+    /// [`CaseFoldExact`](crate::automaton::CaseFoldExact) for the matching model.
     pub fn case_insensitive_iter(&self, needle: &str) -> iter::CaseInsensitiveIter<'_, Data> {
         iter::CaseInsensitiveIter::new(&self.inner, needle)
     }
 
     /// Yield every entry whose case-folded key is within Levenshtein distance
     /// `max_dist` (in codepoints) of `needle`, in lexicographical key order.
-    /// See [`CaseFoldLevenshtein`](automaton::CaseFoldLevenshtein) for the
+    /// See [`CaseFoldLevenshtein`](crate::automaton::CaseFoldLevenshtein) for the
     /// matching model.
     pub fn fuzzy_iter(&self, needle: &str, max_dist: u32) -> iter::FuzzyIter<'_, Data> {
         iter::FuzzyIter::new(&self.inner, needle, max_dist)
@@ -158,7 +157,7 @@ impl<Data> StrTrieMap<Data> {
     /// Codepoint semantics: `?` matches one codepoint (`entr?` matches
     /// `entré`), `*` any run of codepoints. Differs from
     /// [`TrieMap::wildcard_iter`], which matches raw bytes. Matching is
-    /// case-sensitive. See [`CodepointWildcard`](automaton::CodepointWildcard)
+    /// case-sensitive. See [`CodepointWildcard`](crate::automaton::CodepointWildcard)
     /// for the matching model.
     pub fn wildcard_iter(&self, pattern: &str) -> iter::WildcardIter<'_, Data> {
         iter::WildcardIter::new(&self.inner, pattern)
