+++ src/redisearch_rs/trie_rs/src/trie_map/iter/mod.rs
@@ -8,8 +8,16 @@
 */
 
 //! Different iterators to traverse a [`TrieMap`](crate::TrieMap).
-pub mod automaton;
+//!
+//! Filtered traversals come in two flavors. A
+//! [`TraversalFilter`](filter::TraversalFilter) materialises each
+//! candidate key and re-tests it from scratch, paying per key. An
+//! [`Automaton`](crate::automaton::Automaton), driven by
+//! [`AutomatonIter`], is instead stepped incrementally along the trie's
+//! edges, so the state computed for a shared prefix is reused for every
+//! key under it. [`WildcardIter`] dispatches between the two.
 mod contains;
+mod driver;
 pub mod filter;
 mod into_values;
 mod lending;
@@ -22,11 +30,8 @@ mod unfiltered;
 mod values;
 mod wildcard;
 
-pub use automaton::{
-    Automaton, AutomatonIter, AutomatonLendingIter, NfaBitSet, StateClass, WildcardBackend,
-    WildcardIter, WildcardLendingIter, WildcardNfa,
-};
 pub use contains::ContainsIter;
+pub use driver::{AutomatonIter, AutomatonLendingIter};
 pub use into_values::IntoValues;
 pub use lending::LendingIter;
 pub use lending_contains::ContainsLendingIter;
@@ -35,4 +40,6 @@ pub use prefixes::PrefixesIter;
 pub use range::{RangeBoundary, RangeFilter, RangeIter};
 pub use unfiltered::Iter;
 pub use values::Values;
-pub use wildcard::{WildcardFilter, WildcardFilterIter};
+pub use wildcard::{
+    WildcardBackend, WildcardFilter, WildcardFilterIter, WildcardIter, WildcardLendingIter,
+};
